==> content/courses/An Introduction to Biostatistics/03---A-Primer-For-Statistical-Tests.Rmd <==
---
title: "A Primer For Statistical Tests"
subtitle: 'Getting to Terms With Variables'
author: "Erik Kusch"
date: "2020-06-10"
slug: A Primer For Statistical Tests
categories: [An Introduction to Biostatistics]
tags: [R, Statistics]
summary: 'These are the solutions to the exercises contained within the handout to A Primer For Statistical Tests which walks you through the basics of variables, their scales and distributions. Keep in mind that there is probably a myriad of other ways to reach the same conclusions as presented in these solutions.'
authors: [Erik Kusch]
lastmod: '2020-06-10'
featured: no
projects:
output:
  blogdown::html_page:
    toc: true
    toc_depth: 2
    number_sections: false
    fig_width: 6
linktitle: 03 - A Primer For Statistical Tests
menu:
  example:
    parent: Sessions
    weight: 3
# toc: true
type: docs
weight: 3
---
 
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, tidy = TRUE, fig.height = 4.5)
```

# Theory
These are the solutions to the exercises contained within the handout to A Primer For Statistical Tests which walks you through the basics of variables, their scales and distributions. Keep in mind that there is probably a myriad of other ways to reach the same conclusions as presented in these solutions.

I have prepared some [Lecture Slides](/courses/An Introduction to Biostatistics/03---A-Primer-For-Statistical-Tests_Handout.html) for this session.

# Data
Find the data for this exercise [here](/courses/An Introduction to Biostatistics/Data/Primer.RData).

# Loading the `R` Environment Object 
```{r RLoad}
load("Data/Primer.RData") # load data file from Data folder
```

# Variables
## Finding Variables 
```{r RIndentify}
ls() # list all elements in working environment
```
## `Colour`
```{r VColour}
class(Colour) # mode
barplot(table(Colour)) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Colour)` |
| Which scale? | Nominal |
| What's implied? | Categorical data that can't be ordered |
| Does data fit scale? | Yes |



## `Depth` 
```{r VDepth}
class(Depth) # mode
barplot(Depth) # fitting?
```


| Question | Answer |
|-----:|:----|
| Mode? | `r class(Depth)` |
| Which scale? | Interval/Discrete |
| What's implied? | Continuous data with a non-absence point of origin |
| Does data fit scale? | Debatable (is 0 depth absence of depth?) |



## `IndividualsPassingBy` 
```{r VIndPass}
class(IndividualsPassingBy) # mode
barplot(IndividualsPassingBy) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(IndividualsPassingBy)` |
| Which scale? | Integer |
| What's implied? | Only integer numbers with an absence point of origin |
| Does data fit scale? | Yes |



## `Length` 
```{r VLength}
class(Length) # mode
barplot(Length) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Length)` |
| Which scale? | Relation/Ratio |
| What's implied? | Continuous data with an absence point of origin |
| Does data fit scale? | Yes |



## `Reproducing` 
```{r VRepro}
class(Reproducing) # mode
barplot(Reproducing) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Reproducing)` |
| Which scale? | Integer |
| What's implied? | Only integer numbers with an absence point of origin |
| Does data fit scale? | Yes |



## `Sex`
```{r VSex}
class(Sex) # mode
barplot(table(Sex)) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Sex)` |
| Which scale? | Binary |
| What's implied? | Only two possible outcomes |
| Does data fit scale? | Yes |



## `Size` 
```{r VSize}
class(Size) # mode
barplot(table(Size)) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Size)` |
| Which scale? | Ordinal |
| What's implied? | Categorical data that can be ordered |
| Does data fit scale? | Yes |



## `Temperature` 
```{r VTemp}
class(Temperature) # mode
barplot(Temperature) # fitting?
```

| Question | Answer |
|-----:|:----|
| Mode? | `r class(Temperature)` |
| Which scale? | Interval/Discrete |
| What's implied? | Continuous data with a non-absence point of origin |
| Does data fit scale? | Yes (the data is clearly recorded in degree Celsius) |


# Distributions

## `Length`  
```{r DLength}
plot(density(Length)) # distribution plot
shapiro.test(Length) # normality check
```

The data is **normal distributed**.

## `Reproducing` 
```{r DRepro}
plot(density(Reproducing)) # distribution
shapiro.test(Reproducing) # normality check
```

The data is **binomial distributed** (i.e. "How many individuals manage to reproduce") but looks **normal distributed**. The normal distribution doesn't make sense here because it implies continuity whilst the data only comes in integers.

## `IndividualsPassingBy`  
```{r DIndiv}
plot(density(IndividualsPassingBy)) # distribution
shapiro.test(IndividualsPassingBy) # normality check
```

The data is **poisson distributed** (i.e. "How many individuals pass by an observer in a given time frame?").

## `Depth` 
```{r DDepth}
plot(density(Depth)) # distribution
```

The data is **uniform distributed**. You don't know this distribution class from the lectures and I only wanted to confuse you with this to show you that there's much more out there than I can show in our lectures.
