==> content/courses/An Introduction to Biostatistics/12---Simple-Parametric-Tests.Rmd <==
---
title: "Simple Parametric Tests"
subtitle: 'Analysing the Sparrow Data Set'
author: "Erik Kusch"
date: "2020-06-10"
slug: Simple Parametric Tests
categories: [An Introduction to Biostatistics]
tags: [R, Statistics]
summary: 'Welcome to our sixth practical experience in R. Throughout the following notes, I will introduce you to a couple of simple parametric test. Whilst parametric tests are used extremely often in biological statistics, they can be somewhat challenging to fit to your data as you will see soon. To do so, I will enlist the sparrow data set we handled in our first exercise. Additionally, todays seminar is showing plotting via base plot instead of `ggplot2` to highlight the usefulness of base plot and show you the base notation.'
authors: [Erik Kusch]
lastmod: '2020-06-10'
featured: no
projects:
output:
  blogdown::html_page:
    toc: true
    toc_depth: 2
    number_sections: false
    fig_width: 6
linktitle: 12 - Simple Parametric Tests
menu:
  example:
    parent: Sessions 
    weight: 12
# toc: true
type: docs
weight: 12
---

# Theory 
Welcome to our sixth practical experience in R. Throughout the following notes, I will introduce you to a couple of simple parametric test. Whilst parametric tests are used extremely often in biological statistics, they can be somewhat challenging to fit to your data as you will see soon. To do so, I will enlist the sparrow data set we handled in our first exercise. Additionally, todays seminar is showing plotting via base plot instead of `ggplot2` to highlight the usefulness of base plot and show you the base notation.

I have prepared some [Lecture Slides](/courses/An Introduction to Biostatistics/12---Simple-Parametric-Tests_Handout.html) for this session.

# Data
Find the data for this exercise [here](/courses/An Introduction to Biostatistics/Data/1 - Sparrow_Data_READY.rds) and [here](/courses/An Introduction to Biostatistics/Data/2b - Sparrow_ResettledSIUK_READY.rds).

# Preparing Our Procedure
To ensure others can reproduce our analysis we run the following three lines of code at the beginning of our `R` coding file.
```{r Preparation}
rm(list=ls()) # clearing environment
Dir.Base <- getwd() # soft-coding our working directory
Dir.Data <- paste(Dir.Base, "Data", sep="/") # soft-coding our data directory 
```


## Packages
Using the following, user-defined function, we install/load all the necessary packages into our current `R` session.
```{r Packages}
# function to load packages and install them if they haven't been installed yet
install.load.package <- function(x) {
  if (!require(x, character.only = TRUE))
    install.packages(x)
  require(x, character.only = TRUE)
}
package_vec <- c("car") # needed for the Levene Test for Homogeneity
sapply(package_vec, install.load.package)
```

## Loading Data
During our first exercise (Data Mining and Data Handling - Fixing The Sparrow Data Set) we saved our clean data set as an RDS file. To load this, we use the `readRDS()` command that comes with base `R`.
```{r DataLoading}
Data_df_base <- readRDS(file = paste(Dir.Data, "/1 - Sparrow_Data_READY.rds", sep=""))
Data_df <- Data_df_base # duplicate and save initial data on a new object
```


# t-Test (unpaired)

**Assumptions of the unpaired t-Test:**  

- Predictor variable is binary  
- Response variable is metric and **normal distributed** within their groups  
- Variable values are **independent** (not paired)  

In addition, test whether variance of response variable values in groups are equal (`var.test()`) and adjust `t.test()` argument `var.equal` accordingly.

## Testing For Normality And Homogeneity

We need to test the distribution of our response variables within each predictor variable group for their normality and variance. Since this involves two Shapiro tests and one variance test per variable for each response variable, we might want to write our own function to do so:

```{r ShapiroClimate}
ShapiroTest <- function(Variables, Grouping){
  Output <- data.frame(x = Variables)
  for(i in 1:length(Variables)){
    
    X <- Data_df[,Variables[i]]
    Levels <- levels(factor(Data_df[,Grouping]))
    
    Output[i,2] <- shapiro.test(X[which(Data_df[,Grouping] == Levels[1])])$p.value
    Output[i,3] <- shapiro.test(X[which(Data_df[,Grouping] == Levels[2])])$p.value
    Output[i,4] <- var.test(x = X[which(Data_df[,Grouping] == Levels[1])], 
                            y = X[which(Data_df[,Grouping] == Levels[2])])$p.value
  }
  colnames(Output) <- c("Variable", "P.value1", "P.value2", "Var.Test")
  return(Output)
}
```
This function (`ShapiroTest()`) takes two arguments: (1) `Variables` - a vector of characters holding the names of the variables we want to have tested, and (2) `Grouping` - the binary variable by which to group our variables. The function returns a data frame holding the p-values of the Shapiro tests on each variable group values as well as the `var.test()` p-value.  



## Climate Warming/Extremes
**Does sparrow morphology change depend on climate?**

Using multiple different methods (i.e. Kruskal-Wallis and Mann-Whitney U Test), we have already identified climate (be it in its binary form or when recorded as a three-level variable) is a strong driving force of sparrow morphology. We expect the same results when using a t-Test.

Take note that we need to limit our analysis to our climate type testing sites again as follows (we include Manitoba this time as it is at the same latitude as the UK and Siberia and holds a semi-coastal climate type):
```{r Remove2}
# prepare climate type testing data
Data_df <- Data_df_base
Index <- Data_df$Index
Rows <- which(Index == "SI" | Index == "UK" | Index == "RE" | Index == "AU" | Index == "MA")
Data_df <- Data_df[Rows,]
```

### Testing for Normality and Variance
Before we can make use of our data with a t-Test, we need to do an **assumption check**. To this end, we first turn `Climate` records into a binary variable by turning records of a semi-coastal climate into a coastal one.

```{r ShapiroClimate0}
# Make climate binary
Data_df$Climate[which(Data_df$Climate == "Semi-Coastal")] <- "Coastal"
Data_df$Climate <- droplevels(factor(Data_df$Climate))
```

Let's make sure our assumptions are met:
```{r ShapiroCLimate1}
ShapiroTest(Variables = c("Weight", "Height", "Wing.Chord"), Grouping = "Climate")
```
Luckily, all of our variables allow for the calculation of t-Test. Take note though that some need different specification of the `var.equal` argument than others.


### Analyses

**Sparrow Weight**  
Let's start with the weight of *Passer domesticus* individuals as grouped by the climate type present at the site weights have been recorded at:

```{r t-test0a}
t.test(Data_df$Weight ~ Data_df$Climate, var.equal = TRUE)
```
According to our analysis, which has us **reject the null hypothesis**, we conclude that binary climate records are valuable information criteria for predicting sparrow weight with sparrows in coastal climates being lighter than sparrows in continental ones thus effectively varifying the results of our non-parametric approaches (Kruskal-Wallis, Mann-Whitney U).

**Sparrow Height**  
Let's move on to the height of *Passer domesticus* individuals as grouped by the climate type present at the site weights have been recorded at:

```{r t-test0b}
t.test(Data_df$Height ~ Data_df$Climate, var.equal = FALSE)
```
Confirming the results of our Mann-Whitney U Test, we **accept the null hypothesis**.

**Sparrow Wing Chord**  
Lastly, we test the wing chords of *Passer domesticus* individuals as grouped by the climate type present at the site weights have been recorded at:

```{r t-test0c}
t.test(Data_df$Wing.Chord ~ Data_df$Climate, var.equal = FALSE)
```
Without confirming the results of our Mann-Whitney U Test, we **accept the null hypothesis**.



**Conclusion**  
Here's what we've learned from the t-Test so far:  
- Sparrow weight depends on (binary) climate types  
- Sparrow height does not depend on (binary) climate types  
- Sparrow wing chord does not depend on (binary) climate types  

Let's end this by viusalising all of the data:
```{r t-test1, fig.height=12, fig.width=15}
par(mfrow=c(2,2))
plot(Data_df$Weight ~ Data_df$Climate)
plot(Data_df$Height ~ Data_df$Climate)
plot(Data_df$Wing.Chord ~ Data_df$Climate)
```



## Sexual Dimorphism
**Does sparrow morphology change depend on Sex?**

Using the Mann-Whitney U Test, we have already identified the sex of *Passer domesticus* is a good information criterion for understanding sparrow weight but not sparrow height or wing chord. Let's see if we can reproduce this using a t-Test approach.

We may wish to use the entirety of our data set again for this purpose:
```{r over}
Data_df <- Data_df_base
```

### Testing for Normality and Variance
Again, before we can use our data in a t-Test for this purpose, we have to make sure that our assumptions are met. To this end, we can make use of our user defined `ShapiroTest()` function as follows:
```{r ShapiroSex0}
ShapiroTest(Variables = c("Weight", "Height", "Wing.Chord"), Grouping = "Sex")
```
As it turns out, our data does not allow for any t-Test (this happens often in real studies). However, we can create sex-driven subgroups within each site and test whether these meet the requirements for our t-Test. This is out of the scope of this course though and so we will skip it. Spoler alert: I have done this and the findings did not reveal anything we didn't uncover so far.

<!-- In order to do so, we need to do some minor tweaking to our `ShapiroTest()` function: -->
<!-- ```{r ShapiroSex} -->
<!-- ShapiroTestSites <- function(Variables, Grouping){ -->
<!--   list <- list() -->
<!--   for(k in 1:length(unique(Data_df$Index))){ -->
<!--     Output <- data.frame(x = Variables) -->
<!--     Data <- Data_df[which(Data_df$Index == unique(Data_df$Index)[k]), ] -->
<!--     for(i in 1:length(Variables)){ -->

<!--       X <- Data[,Variables[i]] -->
<!--       Levels <- levels(Data[,Grouping]) -->

<!--       Output[i,2] <- shapiro.test(X[which(Data[,Grouping] == Levels[1])])$p.value -->
<!--       Output[i,3] <- shapiro.test(X[which(Data[,Grouping] == Levels[2])])$p.value -->
<!--       Output[i,4] <- var.test(x = X[which(Data_df[,Grouping] == Levels[1])],  -->
<!--                             y = X[which(Data_df[,Grouping] == Levels[2])]) -->
<!--     } -->
<!--     colnames(Output) <- c("Variable", "P.value1", "P.value2", "Var.Test") -->
<!--     list[[k]] <- Output -->
<!--   } -->
<!--   return(list) -->
<!-- } -->
<!-- ``` -->

<!-- This function (`ShapiroTestSites()`) takes two arguments: (1) `Variables` - a vector of characters holding the names of the variables we want to have tested, and (2) `Grouping` - the binary variable by which to group our variables. The function returns a list of data frames for each site holding the p-values of the Shapiro tests on each variable group values as well as the `var.test()` p-value.  -->

<!--  -->

<!-- Let's put this function to the test: -->
<!-- ```{r ShapiroSex1} -->
<!-- ShapiroTestSites(Variables = c("Weight", "Height", "Wing.Chord"), Grouping = "Sex") -->
<!-- ``` -->
<!-- With the exception for sparrow morphology records at:   -->
<!-- - Siberia (`r unique(Data_df$Index)[1]`, height and wing chord of males)   -->
<!-- - Manitoba (`r unique(Data_df$Index)[6]`, morphology of females)   -->
<!-- - South Africa (`r unique(Data_df$Index)[10]`, morphology of females)   -->
<!-- all of our data groups variables are normal distributed with equal variances between the groups per site.   -->

<!-- Since our problematic sites are still relatively close to fulfilling our requirements of the data, we will use them going forward as if they did. -->

<!-- ### Analyses -->

<!-- Running three t-Tests (Weight, Height, Wing Chord) for each of our eleven sites is absolute mania! Therefore, we write our own function again that let's us apply the tests exactly the way we want to: -->
<!-- ```{r ttestSex0} -->
<!-- t_testSite <- function(Variables, Grouping, data, VarEqual){ -->
<!--   Data <- data -->
<!--   Index <- unique(Data$Index) -->
<!--   Indexes <- Data$Index -->
<!--   list <- list() -->
<!--   for(i in 1:length(Variables)){ -->
<!--     Output <- data.frame(NA) -->
<!--     for(k in 1:length(Index)){ -->
<!--       # data and test -->
<!--       X <- Data[, Variables[i]][which(Indexes == Index[k])] -->
<!--       Y <- Data[, Grouping][which(Indexes == Index[k])] -->
<!--       Test <- t.test(X ~ Y, paired = FALSE, var.equal = VarEqual) -->
<!--       # filling data frame -->
<!--       Output[1,k] <- Test[["p.value"]] -->
<!--       Output[2,k] <- Test[["estimate"]][[1]] -->
<!--       Output[3,k] <- Test[["estimate"]][[2]] -->
<!--     } -->
<!--     # data frame to list -->
<!--     colnames(Output) <- Index -->
<!--     rownames(Output) <- c("p", "Mean1", "Mean2") -->
<!--     list[[i]] <- Output -->
<!--   } -->
<!--   return(list) -->
<!-- } -->
<!-- ``` -->

<!-- This `t_testSite()` function takes four arguments: (1) `Variables` - a vector of characters holding the names of the variables we want to have tested, (2) `Grouping` - the binary variable by which to group our variables, (3) `data` - the data frame which contains the `Variables` and the `Grouping` factor, and (4) `VarEqual` - a logical indicator of whether to perform a t-Test assuming equal variance of the groups or not.   -->
<!-- The function returns a list of data frames (one per variable) containing the p-values of the unpaired t-Tests for each variable at every site as well as the predicted group means. -->

<!--  -->

<!-- Although our function `t_testSite()` can handle multiple variables at once, we will now use it on each of our morphological sparrow variables individually to disentangle them a bit easier:   -->

<!-- \begin{center} -->
<!-- \textit{Does sparrow weight depend on sex when assessed at each of our sites individually?} -->
<!-- \end{center} -->
<!-- ```{r tTestSex1} -->
<!-- t_testSite(Variables = "Weight", Grouping = "Sex", data = Data_df, VarEqual = TRUE) -->
<!-- ``` -->
<!-- As it turns out, sex is a statistically significant predictor for sparrow weight at each site. This was to be expected. Using a binomial test in our second practical, we identified no bias in sexes for our sparrow populations. In addition, using the Mann-Whitney U-Test in our fourth practical, we identified sex to be an important information criterion for sparrow weight across all of our sites. Given these two conditions, we were expecting a results like the one presented here with males being, on average, heavier than females in *Passer domesticus* and we **reject the null hypothesis**.   -->

<!-- \hfill \linebreak -->
<!-- \begin{center} -->
<!-- \textit{Does sparrow height depend on sex when assessed at each of our sites individually?} -->
<!-- \end{center} -->
<!-- ```{r tTestSex2} -->
<!-- t_testSite(Variables = "Height", Grouping = "Sex", data = Data_df, VarEqual = TRUE) -->
<!-- ``` -->
<!-- Like with our Mann-Whitney U-Test, we fail to identify a significant effect of sex on sparrow height records at each of our sites and so we **accept the null hypothesis**. -->

<!-- \hfill \linebreak -->
<!-- \begin{center} -->
<!-- \textit{Does sparrow wing chord depend on sex when assessed at each of our sites individually?} -->
<!-- \end{center} -->
<!-- ```{r tTestSex3} -->
<!-- t_testSite(Variables = "Wing.Chord", Grouping = "Sex", data = Data_df, VarEqual = TRUE) -->
<!-- ``` -->
<!-- Like with our Mann-Whitney U-Test, we fail to identify a significant effect of sex on sparrow wing chord records at each of our sites and so we **accept the null hypothesis**. -->

<!--  -->

# t-Test (paired)

**Assumptions of the paired t-Test:**  

- Predictor variable is binary  
- Response variable is metric  
- *Difference of response variable pairs* is **normal distributed**  
- Variable values are **dependent** (paired)  

## Preparing Data
For this purpose, we need an **additional data set with truly paired records** of sparrows and so we implement the same solution as we've used within our fourth seminar using the Wilcoxon Signed Rank Test. Within our study set-up, think of a **resettling experiment**, were you take *Passer domesticus* individuals from one site, transfer them to another and check back with them after some time has passed to see whether some of their characteristics have changed in their expression.  
To this end, presume we have taken the entire *Passer domesticus* population found at our **Siberian** research station and moved them to the **United Kingdom**. Whilst this keeps the latitude stable, the sparrows *now experience a coastal climate instead of a continental one*. After some time (let's say: a year), we have come back and recorded all the characteristics for the same individuals again.  

You will find the corresponding *new data* in `2b - Sparrow_ResettledSIUK_READY.rds`. Take note that this set only contains records for the transferred individuals in the **same order** as in the old data set.  

```{r DataLoading2}
Data_df_Resettled <- readRDS(file = paste(Dir.Data, "/2b - Sparrow_ResettledSIUK_READY.rds", sep=""))
```

Since earlier analysis such as the Wilcoxon Signed Rank test (fourth practical) and the Friedman Test (fifth practical) showed that height and wing chord records do not change when sparrows are resettled at all, we have excluded these here and **focus solely on sparrow weight**.

## Testing for Normality
Before being able to run our paired t-Test, we must make sure that the *difference of response variable pairs* is **normal distributed**. We can do so using the `shapiro.test()` of base `R` as follows:
```{r Norm2}
# selecting pre-resettling weights
DataSI <- Data_df$Weight[which(Data_df$Index == "SI")]
# calculating difference of before and after resettling weights
WeightDiff <- DataSI-Data_df_Resettled$Weight
# shapiro test
shapiro.test(WeightDiff)
```
Thankfully, the **assumption of normality** is **met**.

Now let's visualise that using a qqplot:
```{r Norm3, fig.height = 8}
qqnorm(WeightDiff)
qqline(WeightDiff)
```

## Climate Warming/Extremes
**Does sparrow morphology change depend on climate?**

Now let's go on to test whether sparrow weights change significantly per individual due to our relocation experiment (we expect this from future test in our practicals):
```{r tpaired}
t.test(DataSI, Data_df_Resettled$Weight, paired = TRUE)
```
We were right, individual sparrow weights change significantly after our relocation experiment and we **reject the null hypothesis**. This is in accordance with the results of the Wilcoxon Signed Rank Test as well as the Friedman Test.


Let's go on to visualise our data to make better sense of what is going on here:
```{r tpaired1, fig.height=8}
# Select the sparrow weights
Weights <- c(DataSI, Data_df_Resettled$Weight)
# Select the sites
Sites <- factor(rep(c("SI", "SI_UK"), each = length(DataSI)))
# Plot
plot(Weights ~ Sites)
```

Quite obviously sparrows observed in Siberia are heavier than when they are resettled to the United Kingdom (this may be due to the more forgiving climate in the UK). Just like the test stated, the difference of the average weights is roughly 2g between the sparrows at the two sites.


# One-Way ANOVA
**Assumptions of the One-Way ANOVA:**  

- Predictor variable is categorical  
- Response variable is metric  
- *Response variable residuals* are **normal distributed**  
- Variance of populations/samples are equal (**homogeneity**)  
- Variable values are **independent** (not paired)  

## Testing For Assumptions

Firstly, we need to test the assumptions of our One-Way ANOVA. For this purpose, we write another user-defined function.
```{r AssCheck1}
# User-defined function
ANOVACheck <- function(Variables, Grouping, data, plotting){
  Output <- data.frame(x = NA)
  for(i in 1:length(Variables)){
    # data
    Y <- as.numeric(factor(data[,Variables[i]]))
    X <- data[,Grouping]
    Levels <- levels(factor(Data_df[,Grouping]))
    # Residuals?
    model <- lm(Y ~ X)
    Output[1,i] <- shapiro.test(residuals(model))$p.value
    # Homgeneity?
    Levene <- leveneTest(Y ~ X, 
                         center = median, 
                         data = data)
    Output[2,i] <- Levene[1,3]
    # Plotting
    if(plotting == TRUE){
      plot(model, 2)# Normality
      plot(model, 3)# Homogeneity
    }
  }
  colnames(Output) <- Variables
  rownames(Output) <- c("Residual Normality", "Homogeneity of Variances")
  return(Output)
}
```

This `ANOVACheck()` function takes four arguments: (1) `Variables` - a vector of characters holding the names of the variables we want to have tested, (2) `Grouping` - the categorical variable by which to group our variables, (3) `data` - the data frame which contains the `Variables` and the `Grouping` factor, and (4) `plotting` - a logical indicator of whether to produce plots visualising the test results or not.  
The function returns a data frames containing the p-values indexing whether to accept or reject the notion of the normality of residuals per variable (`Residual Normality`), and the p-values indexing whether variances between groups are homogeneous or not (`Homogeneity of Variances`).  



## Climate Warming/Extremes
**Does sparrow morphology change depend on climate?**

Using the Kruskal-Wallis Test in our last exercise, we already identified climate to be an important factor in determining *Passer domesticus* morphology. Let's see if this holds true.

Take note that we need to limit our analysis to our climate type testing sites again as follows (we include Manitoba this time as it is at the same latitude as the UK and Siberia and holds a semi-coastal climate type):
```{r Remove3}
# prepare climate type testing data
Data_df <- Data_df_base
Index <- Data_df$Index
Rows <- which(Index == "SI" | Index == "UK" | Index == "RE" | Index == "AU" | Index == "MA")
Data_df <- Data_df[Rows,]
```

### Assumption Check
Let's use the `ANOVACheck()` function on our data:
```{r AssCheck1b, fig.height = 8}
par(mfrow=c(3,2))
ANOVACheck(Variables = c("Weight", "Height", "Wing.Chord"), Grouping = "Climate", data = Data_df, plotting = TRUE)
```

Unfortunately, neither weight nor wing chord records fullfil our requirements.



### Analysis
Let's run our analysis for height as grouped by the three-level climate variable:
```{r Analysis1a}
model <- lm(Data_df$Height ~ Data_df$Climate)
anova(model)
```
According to this, climate is a meaningful predictor of height of sparrows and we **reject the null hypothesis** thus confirming the results of our Kruskall-Wallis analysis.  

Now, let's analyse the output a bit more in-depth:
```{r Analysis1b}
summary(model)
```

- The mean sparrow height in coastal climates is `r model$coefficients[1]`cm (this is our **Intercept**/*Baseline*)  
- The mean sparrow height in continental climates is `r model$coefficients[2]`cm bigger than the **Intercept**  
- The mean sparrow height in semi-coastal climates is `r model$coefficients[3]`cm bigger than the **Intercept**
- Only the estimates in coastal and semi-coastal climates are statistically significant  

Personally, I would not place too much confidence in these results due to a couple of reasons:  

- Our only semi-coastal site is on the northern hemisphere whereas two of our stations are located in the southern hemisphere  
- Confounding factors such as population status might have an effect which we are not considering here  



Let's end this by plotting all of our data:
```{r Analysis1, fig.height=12}
par(mfrow=c(2,2))
plot(Data_df$Weight ~ factor(Data_df$Climate))
plot(Data_df$Height ~ factor(Data_df$Climate))
plot(Data_df$Wing.Chord ~ factor(Data_df$Climate))
```

As you can see, the variances are definitely not equal between our groups which explains why part of our assumption test failed.



## Predation
**Does nesting height depend on predator characteristics?**

Again, using the Kruskal-Wallis Test in our last exercise, we already identified predator characteristics to be an important factor in determining *Passer domesticus* nesting height. Let's see if this holds true.

We may wish to use the entirety of our data set again for this purpose:
```{r over2}
Data_df <- Data_df_base
```

### Assumption Check

Let's use our `ANOVACeck()` function to test whether we can run our analysis. Before we can do so, however, we need to slightly adjust our predator type variable just like we did in our last exercise and as follows:
```{r AssCheck2, fig.height = 7}
# changing levels in predator type
levels(Data_df$Predator.Type) <- c(levels(Data_df$Predator.Type), "None")
Data_df$Predator.Type[which(is.na(Data_df$Predator.Type))] <- "None"

# Assumption Check
par(mfrow=c(1,2))
ANOVACheck(Variables = "Nesting.Height", Grouping = "Predator.Type", data = Data_df, plotting = TRUE)
```
Again, our data fails the assumption check. The residuals are definitely not normal distributed and the variance of nesting height records within our groups are not equal.



### Analysis

Since none of our assumptions are met, we cannot run an ANOVA and therefore resort to data visualisation alone:
```{r Analysis2, fig.height=12}
plot(Data_df$Nesting.Height ~ Data_df$Predator.Type)
```
Once more, we can see why our homogeneity of variances test failed.

# Two-Way ANOVA

**Assumptions of the Two-Way ANOVA:**  

- Predictor variables are categorical  
- Response variable is metric  
- *Response variable residuals* are **normal distributed**  
- Variance of populations/samples are equal (**homogeneity**)  
- Variable values are **independent** (not paired)  

## Testing For Assumptions

Yet again, we need to check if our assumptions are met first. Automating this procedure is definitely a good idea and only needs slight modification from our `ANOVACheck()` function.
```{r AssCheck4}
# User-defined function
ANOVACheck_TWO <- function(Formulas, data, plotting){
  Output <- data.frame(x = NA)
  for(i in 1:length(Formulas)){
    # Check how many formulas there are
    if(length(Formulas) == 1){
      Expression <- Formulas[[1]]
    }else{
      Expression <- Formulas[[i]]
    }
    # Residuals?
    model <- lm(formula = Expression, data = data)
    Output[1,i] <- shapiro.test(residuals(model))$p.value
    # Homgeneity?
    Levene <- leveneTest(Expression, 
                         center = median, 
                         data = data)
    Output[2,i] <- Levene[1,3]
    # Plotting
    if(plotting == TRUE){
      plot(model, 2)# Normality
      plot(model, 3)# Homogeneity
    }
  }
  colnames(Output) <- as.character(Formulas)
  rownames(Output) <- c("RN", "HoV")
  return(Output)
}
```

This `ANOVACheck_TWO()` function takes four arguments: (1) `Formulas` - a vector of formula specification for our ANOVA models we want to have tested, (2) `data` - the data frame which contains the variables and the grouping factor called upon in our `Formulas`, and (3) `plotting` - a logical indicator of whether to produce plots visualising the test results or not.  
The function returns a data frames containing the p-values indexing whether to accept or reject the notion of the normality of residuals per variable (`RN`), and the p-values indexing whether variances between groups are homogeneous or not (`HoV`).

## Sexual Dimorphism
**Does sparrow morphology depend on population status and sex?**

Given different factors affecting invasive species, we might expect different patterns of sexual dimorphism for invasive and native populations. Take note that we keep using the northern hemisphere subset our cimate testing sites as these present us with a nice set of invasive/native population records already whilst keeping confounding factors to a minimum.

### Assumption Check
First, we need to check our assumptions:
```{r AssCheck5, fig.height = 6.5}
# prepare climate type testing data
Data_df <- Data_df_base
Index <- Data_df$Index
Rows <- which(Index == "SI" | Index == "UK" | Index == "MA")
Data_df <- Data_df[Rows,]

# analysis
par(mfrow=c(3,2))
ANOVACheck_TWO(Formulas = c(Weight ~ Population.Status*Sex, 
                            Height ~ Population.Status*Sex,
                            Wing.Chord ~ Population.Status*Sex)
               , data = Data_df, plotting = TRUE)
```
Again our assumptions are not met except for sparrow height and wing chord as a product of sex and population status.


### Analysis
Let's run our analysis:
```{r Analysis5, fig.height = 6}
# height model
model <- lm(Height ~ Population.Status*Sex, data = Data_df)
anova(model)
# wing chord model
model <- lm(Wing.Chord ~ Population.Status*Sex, data = Data_df)
anova(model)
# plotting
par(mfrow=c(1,2))
boxplot(Height ~ Population.Status*Sex, data = Data_df, col = c("blue", "red"))
boxplot(Wing.Chord ~ Population.Status*Sex, data = Data_df, col = c("blue", "red"))
```
As it turns out, population status and sex are no viable predictors for sparrow height or wing chord and so we **accept the null hypothesis**.


# ANCOVA

**Assumptions of the ANCOVA:**

- Predictor variables are categorical or continuous  
- Response variable is metric  
- *Response variable residuals* are **normal distributed**  
- Variance of populations/samples are equal (**homogeneity**)  
- Variable values are **independent** (not paired)  
- Relationship between the response and covariate is **linear**.  

## Climate Warming/Extremes
**Do sparrow characteristics depend on climate and latitude?**

Latitude may have masked some effects of climate on sparrow morphology in our preceding analyses and vice-versa. At times, we have been able to account for this by including our site records, which can be seen as binned versions of latitude records. Let's test if the inclusion of raw latitude records are meaningful.

### Assumption Check
Again, we need to do an assumption check. However, we need a new function for this, since we now need to test whether our response variable and the covariate are linear or not:
```{r AssCheck6}
# overwriting prior changes in Data_df
Data_df <- Data_df_base
Data_df$Latitude <- abs(Data_df$Latitude)
# User-defined function
ANCOVACheck <- function(Variables, Grouping, Covariate, data, plotting){
  Output <- data.frame(x = NA)
  for(i in 1:length(Variables)){
    # data
    Y <- as.numeric(factor(data[,Variables[i]]))
    X <- factor(data[,Grouping])
    Z <- data[, Covariate]
    # Residuals?
    model <- lm(Y ~ X*Z)
    Output[1,i] <- shapiro.test(residuals(model))$p.value
    # Homgeneity?
    Levene <- leveneTest(Y ~ X, 
                         center = median, 
                         data = data)
    Output[2,i] <- Levene[1,3]
    # Plotting
    if(plotting == TRUE){
      plot(model, 1)# Linearity
      plot(model, 2)# Normality
      plot(model, 3)# Homogeneity
    }
  }
  colnames(Output) <- Variables
  rownames(Output) <- c("RN", "HoV")
  return(Output)
}
```
This `ANCOVACheck()` function takes five arguments: (1) `Variables` - a vector of response variables used in our models, (2) `Grouping` - the categorical variable by which to group our variables, (3) `Covariate` - the covariate of our analysis, (4)`data` - the data frame which contains the variables, the grouping factor and our covariate, and (5) `plotting` - a logical indicator of whether to produce plots visualising the test results or not.  
The function returns a data frames containing the p-values indexing whether to accept or reject the notion of the normality of residuals per variable (`RN`), and the p-values indexing whether variances between groups are homogeneous or not (`HoV`).

```{r AssCheck6b, fig.height=10}
ANCOVACheck(Variables = c("Weight", "Height", "Wing.Chord", "Nesting.Height", "Egg.Weight", "Number.of.Eggs", "Home.Range"), 
            Grouping = "Climate", Covariate = "Latitude", 
            data = Data_df, plotting = FALSE)
```
The assumptions aren't met. I have set the `plotting` argument to `FALSE` tu suppress the plotting of model checking visualisation. The would be useful to judge linearity but not necessary here since the other two important assumptions (Homogeneity of variances and Normality of residuals) aren't met to begin with.

### Analysis
Since none of our assumptions are met, we cannot run an ANOVA and therefore resort to data visualisation alone. We need a new function for this to do our plotting easily and automatically with some colours indicating our grouping factors whilst plotting response variables versus covariates.
```{r Analysis6}
PlotAncovas <- function(Variables, Grouping, Covariate, data){
  for(i in 1:length(Variables)){
    Y <- Data_df[,Variables[i]]
    if(class(Y) == "character"){Y <- factor(Y)}
    X <- Data_df[,Covariate]
    G <- factor(Data_df[, Grouping])
    plot(X, Y, col = G, xlab = Covariate, ylab = Variables[i])
    legend("top", # place legend at the top
           inset = -0.35, # move legend away from plot centre
           xpd = TRUE, # allow legend outside of plot area
           legend=levels(G), # what to include in legend
           bg = "white", col = unique(G), ncol=length(levels(G)), # colours
           pch = 1, # plotting symbols
           title = Variables[i] # title of legend
           )
  }
}
```
The `PlotAncovas()` returns a scatter plot and takes four arguments: (1) `Variables` - a vector of response variables, (2) `Grouping` - the name of the grouping factor according to which to colour the symbols in our plot, (3) `Covariate` - the covariate against which to plot individuals variables, and (4) `data` - the data frame which holds our variables.  

Let's use our function:
```{r Analysisb, fig.height=4, eval=FALSE}
par(mfrow=c(1,2))
PlotAncovas(Variables = c("Weight", "Height", "Wing.Chord", "Nesting.Height", "Egg.Weight", "Number.of.Eggs", "Home.Range"), Grouping = "Climate", Covariate = "Latitude", data = Data_df)
```

I will not interpret these plots here in text and leave this to you.  

Take note that this **could've been achieved much easier with `ggplot2`!**


```{r Analysisc, fig.height=4, echo=FALSE}
par(mfrow=c(1,2))
PlotAncovas(Variables = c("Weight", "Height", "Wing.Chord", "Nesting.Height", "Egg.Weight", "Number.of.Eggs", "Home.Range"), Grouping = "Climate", Covariate = "Latitude", data = Data_df)
```




## Sparrow Characteristics And Sites

*This was not part of what we set out to do according to the lecture slides but has been included as a logical conclusion to an earlier analysis.*

Unfortunately, our previous attempt at an ANCOVA didn't work. So what other covariate do we have available for sparrow characteristics?  

- *Latitude* doesn't make sense to include when grouping by site index as these two are synonymous  
- *Longitude* doesn't make sense to include when grouping by site index as these two are synonymous  
- *Weight* is well explained by other variables and we know the causal links  
- *Height* is not that well explained by other variables  
- *Wing.Chord* is not that well explained by other variables  

Of course, there are more within our data set but it has become apparent that `Weight` may make for an important covariate in our site-wise ANCOVA set-up. Using the Pearson correlation (third practical), we already identified a causal link between sparrow `Weight` and `Height` per site.

### Assumption Check
Firstly, we test whether assumptions are met. For brevities sake, we only test four variables:
```{r AssCheck7, fig.height=3}
par(mfrow=c(1,3))
ANCOVACheck(Variables = c("Height", "Wing.Chord", "Egg.Weight", "Number.of.Eggs"), Grouping = "Index", Covariate = "Weight", data = Data_df, plotting = TRUE)
```
As it turns out, we can run our ANCOVA on `Egg.Weight` when grouped by site `Index` and driven by `Weight`.

### Analysis
First, let's visualise our data:
```{r Analysis7, fig.height=4}
PlotAncovas(Variables = "Egg.Weight", Grouping = "Index", Covariate = "Weight", data = Data_df)
```

Quite obviously, Belize (BE) records are very different from the other stations, whose egg weight and weight records are grouped together. There seems to be some evidence for an overall linkage of sparrow weight and egg weight (a positive correlation).

Now we run the analysis:
```{r Analysis7b}
LM_fit5 <- lm(Egg.Weight ~ Weight*Index, data = Data_df)
anova(LM_fit5)
```
The above ANCOVA output tells us that there is no interaction effect between sites and sparrow weights when determining mean egg weight per nest of *Passer domesticus* and so we do another iteration of our model and remove the postulated interaction:
```{r Analysis7c}
LM_fit6 <- lm(Egg.Weight ~ Weight+Index, data = Data_df)
anova(LM_fit6)
```

By now, all of our model coefficients are significant and we can go on to interpret them:
```{r Analysis7d}
summary(LM_fit6)
```
